test: remove unnecessary tests
