+++ tests/test_topic_create_transaction.py
@@ -1,13 +1,12 @@
 """Tests for the TopicCreateTransaction functionality."""
 
 import pytest
-from unittest.mock import patch
+
 from hiero_sdk_python.consensus.topic_create_transaction import TopicCreateTransaction
 from hiero_sdk_python.account.account_id import AccountId
 from hiero_sdk_python.crypto.private_key import PrivateKey
 from hiero_sdk_python.response_code import ResponseCode
 from hiero_sdk_python.consensus.topic_id import TopicId
-from hiero_sdk_python.exceptions import PrecheckError
 from hiero_sdk_python.hapi.services import (
     basic_types_pb2,
     response_header_pb2,
@@ -121,81 +120,3 @@ def test_execute_topic_create_transaction():
         assert receipt.topicId.shard == 0
         assert receipt.topicId.realm == 0
         assert receipt.topicId.num == 123
-
-def test_topic_create_transaction_retry_on_busy():
-    """Test that TopicCreateTransaction retries on BUSY response."""
-    # First response is BUSY, second is OK
-    busy_response = transaction_response_pb2.TransactionResponse(
-        nodeTransactionPrecheckCode=ResponseCode.BUSY
-    )
-    
-    ok_response = transaction_response_pb2.TransactionResponse(
-        nodeTransactionPrecheckCode=ResponseCode.OK
-    )
-    
-    receipt_response = response_pb2.Response(
-        transactionGetReceipt=transaction_get_receipt_pb2.TransactionGetReceiptResponse(
-            header=response_header_pb2.ResponseHeader(
-                nodeTransactionPrecheckCode=ResponseCode.OK
-            ),
-            receipt=transaction_receipt_pb2.TransactionReceipt(
-                status=ResponseCode.SUCCESS,
-                topicID=basic_types_pb2.TopicID(
-                    shardNum=0,
-                    realmNum=0,
-                    topicNum=456
-                )
-            )
-        )
-    )
-    
-    response_sequences = [
-        [busy_response, ok_response, receipt_response],
-    ]
-    
-    with mock_hedera_servers(response_sequences) as client, patch('time.sleep') as mock_sleep:
-        client.max_attempts = 3
-        
-        tx = (
-            TopicCreateTransaction()
-            .set_memo("Test with retry")
-            .set_admin_key(PrivateKey.generate().public_key())
-        )
-        
-        try:
-            receipt = tx.execute(client)
-        except Exception as e:
-            pytest.fail(f"Should not raise exception, but raised: {e}")
-        # Verify transaction succeeded after retry
-        assert receipt.status == ResponseCode.SUCCESS
-        assert receipt.topicId.num == 456
-        
-        # Verify we slept once for the retry
-        assert mock_sleep.call_count == 1, "Should have retried once"
-        
-        # Verify we didn't switch nodes (BUSY is retriable without node switch)
-        assert client.node_account_id == AccountId(0, 0, 3), "Should not have switched nodes on BUSY"
-
-def test_topic_create_transaction_fails_on_nonretriable_error():
-    """Test that TopicCreateTransaction fails on non-retriable error."""
-    # Create a response with a non-retriable error
-    error_response = transaction_response_pb2.TransactionResponse(
-        nodeTransactionPrecheckCode=ResponseCode.INVALID_TRANSACTION_BODY
-    )
-    
-    response_sequences = [
-        [error_response],
-    ]
-    
-    with mock_hedera_servers(response_sequences) as client, patch('time.sleep'):
-        tx = (
-            TopicCreateTransaction()
-            .set_memo("Test with error")
-            .set_admin_key(PrivateKey.generate().public_key())
-        )
-        
-        with pytest.raises(PrecheckError) as exc_info:
-            tx.execute(client)
-        
-        # Verify the error contains the expected status
-        assert str(ResponseCode.INVALID_TRANSACTION_BODY) in str(exc_info.value)
